// src/types/timetable.ts
export type DayOfWeek = '월' | '화' | '수' | '목' | '금';

// request
export interface TimetableForm {
    year: number;
    semester: number;
    name: string;
    dayPeriod: string;
    credit: number;
    code: string;
}

export interface TimetableAutoGeneratedForm { // 시간표 자동 생성 시 폼
    year: number;
    semester: number;
    targetCredits: number; // 목표 학점
    geCategories: Array<string>; // 교양분야
    freeDays: Array<string>; // 공강
}

// response
export interface TimeTableDto { // api의 결과로 얻는 데이터
    timetableId: number;
    year: number;
    semester: number;
    totalCredits: number;
    lectures: Array<Lecture>;
}

export interface Lecture {
    itemId: number;
    lectureId: number;
    userLectureId: number | null;
    fixed: boolean;
    code: string;
    name: string;
    professor: string;
    dayPeriod: string; // 시간(월25.26)
    classroom: string; // 장소(혜-409)
    geCategory: string;
    credit: number;
}

// 화면 렌더링 용 데이터(파싱해서 사용해야함!)
export interface Schedule {
    id: number;
    day: DayOfWeek;
    startTime: number;
    endTime: number;
    name: string;
    location?: string;
    professor?: string;
}